# openapi.yaml
openapi: 3.0.3
info:
  title: Mini-Arke - Manufacturing Management API
  version: 1.0.0
  description: API for managing products, orders, and sales analytics

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /products:
    get:
      summary: List all products
      operationId: listProducts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /products/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get a product by ID
      operationId: getProduct
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update a product
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete a product
      operationId: deleteProduct
      responses:
        '204':
          description: Product deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - sku
        - name
        - price
        - stockQuantity
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          maxLength: 100
        name:
          type: string
          maxLength: 200
        description:
          type: string
          nullable: true
        price:
          type: number
          format: float
          minimum: 0
        stockQuantity:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true

    CreateProductRequest:
      type: object
      required:
        - sku
        - name
        - price
      properties:
        sku:
          type: string
          maxLength: 100
        name:
          type: string
          maxLength: 200
        description:
          type: string
        price:
          type: number
          format: float
          minimum: 0
        stockQuantity:
          type: integer
          minimum: 0
          default: 0

    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
        price:
          type: number
          format: float
          minimum: 0
        stockQuantity:
          type: integer
          minimum: 0

    ProductList:
      type: object
      required:
        - products
        - total
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer

    MonthlySalesStats:
      type: object
      required:
        - months
      properties:
        months:
          type: array
          items:
            $ref: '#/components/schemas/MonthlyStat'

    MonthlyStat:
      type: object
      required:
        - month
        - totalRevenue
        - totalOrders
        - averageOrderValue
      properties:
        month:
          type: string
          description: Month in format YYYY-MM
          example: "2024-11"
          pattern: '^\d{4}-\d{2}$'
        totalRevenue:
          type: number
          format: float
          description: Total revenue for the month
          minimum: 0
        totalOrders:
          type: integer
          description: Total number of orders
          minimum: 0
        averageOrderValue:
          type: number
          format: float
          description: Average order value
          minimum: 0

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: string
          nullable: true

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'